Core/Game: Fix remaining warnings on windows (and most likely almost all of the ones on *nix)
--HG--
branch : trunk

+++ src/server/game/Accounts/AccountMgr.h
@@ -51,7 +51,7 @@ class AccountMgr
 
         uint32 GetId(std::string username);
         uint32 GetSecurity(uint32 acc_id);
-        uint32 GetSecurity(uint32 acc_id, int32 realm_id);
+        uint32 GetSecurity(uint64 acc_id, int32 realm_id);
         bool GetName(uint32 acc_id, std::string &name);
         uint32 GetCharactersCount(uint32 acc_id);
         std::string CalculateShaPassHash(std::string& name, std::string& password);
