+++ src/server/game/Server/Protocol/Handlers/PetHandler.cpp
@@ -614,7 +614,7 @@ void WorldSession::HandlePetRename(WorldPacket & recv_data)
     trans->PAppend("UPDATE character_pet SET name = '%s', renamed = '1' WHERE owner = '%u' AND id = '%u'", name.c_str(), _player->GetGUIDLow(), pet->GetCharmInfo()->GetPetNumber());
     CharacterDatabase.CommitTransaction(trans);
 
-    pet->SetUInt32Value(UNIT_FIELD_PET_NAME_TIMESTAMP, time(NULL));
+    pet->SetUInt32Value(UNIT_FIELD_PET_NAME_TIMESTAMP, uint32(time(NULL))); // cast can't be helped
 }
 
 void WorldSession::HandlePetAbandon(WorldPacket & recv_data)
